+++ templates/homepage.html
@@ -133,29 +133,22 @@
               All Libraries&nbsp;<i class="fas fa-chevron-right  text-sky-600 dark:text-sky-300 group-hover:text-orange dark:group-hover:text-orange"></i>
             </a>
           </div>
-          <h3 class="pb-2 mb-4 text-lg md:text-2xl capitalize border-b border-gray-400 text-orange dark:border-slate"><a href="{% url 'library-detail' library_slug=featured_library.slug version_slug=LATEST_RELEASE_URL_PATH_STR %}" class="link-header">{{ featured_library.name }}</a></h3>
+          <h3 class="pb-2 mb-4 text-lg md:text-2xl capitalize border-b border-gray-400 text-orange dark:border-slate"><a href="{% url 'library-detail' library_slug=featured_library.library.slug version_slug=LATEST_RELEASE_URL_PATH_STR %}" class="link-header">{{ featured_library.library.name }}</a></h3>
           <span class="pb-1 mx-auto w-full text-sm md:text-base align-left">
             {{ featured_library.description }}
           </span>
           <p class="pb-1 mx-auto w-full text-xs md:text-sm align-left">{% if featured_library.first_boost_version %}Added in {{ featured_library.first_boost_version.display_name }}{% endif %}</p>
           <div class="py-4">
-            {% if featured_library.authors %}
+            {% if authors or maintainers %}
               <div class="flex flex-wrap justify-start">
-              {% for author in featured_library.authors.all %}
-                <div class="p-1 md:p-2 w-min  text-center flex flex-col items-center justify-center">
-                  <div class="bg-gray-300 dark:bg-slate rounded-lg w-[36px] h-[36px]">
-                  {% if author.image %}
-                    <img src="{{ author.image.url }}"
-                      title="{{ author.display_name }}"
-                      alt="{{ author.display_name }}"
-                      class="rounded-lg w-[36px] h-[36px]" />
-                  {% else %}
-                    <i class="text-3xl fas fa-user text-white dark:text-white/60" title="{{ author.display_name }}"></i>
-                  {% endif %}
-                  </div>
-                  <span class="text-xs">{{ author.display_name }}</span>
+                <div class="grid md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-y-3 gap-x-2">
+                  {% for user in authors %}
+                    {% avatar user=user commitauthor=user.commitauthor avatar_type="wide" contributor_label="Author" %}
+                  {% endfor %}
+                  {% for user in maintainers %}
+                    {% avatar user=user commitauthor=user.commitauthor avatar_type="wide" contributor_label="Maintainer" %}
+                  {% endfor %}
                 </div>
-              {% endfor %}
               </div>
             {% endif %}
